Properly handle read-only clusters in m_pullup(9).
If the first mbuf of a chain in m_pullup is a cluster, check if the
cluster is read-only (shared or an external buffer).  If so, don't
touch it and create a new mbuf for the pullup data.

This restores original 4.4BSD m_pullup, that not only returned
contiguous mbuf data of the specified length, but also converted
read-only clusters into writeable memory.  The latter feature was
lost during some refactoring.

from ehrhardt@; tested by weerd@; OK stsp@ bluhm@ claudio@