add support for "rockchip,cryptov2-rng"
the steps to operate rockchip,cryptov2-rng are basically the same
as the existing rockchip,cryptov1-rng support, but the registers
and bits have moved around. add some abstraction for the register
differences and have the state machine call the different backends.

this is present on rk356x chips as the "True Random Number Generator
(TRNG)".

tested on a bunch of different rk3568 boards.
ok kettenis@