clockintr_cpu_init: avoid CQ_INIT flag when scheduling cq_hardclock
The meaning of the CQ_INIT flag is about to change.  Soon, we won't be
able to use it to decide whether a given clockintr_cpu_init() call is
the first on a given CPU.

Instead, use the value of cl_expiration.  If it's zero, we know this
is the first clockintr_cpu_init() call on this CPU.