Add some RK3588 PCIe related clocks.
Also add some RK3588 resets.  Whoever reviewed the bindings on Linux gave
the brilliant advice that clock IDs and reset IDs should not in any way
have a sane mapping to the hardware registers, even though that is the
case on all older Rockchip SoCs and greatly simplifies the driver.  So
now we need to implement pointless lookup code.

ok patrick@, dlg@