Revert rev 1.129: When detaching devices when we suspend, we need to continue processing command completion events.  So only return early in xhci_softintr() if the controller is dead instead of dying.  This fixes USB suspend/resume in Apple M1/M2.
ok mlarkin@, deraadt@
-----

The change does not only allow completion events to be processed but also
events that get processed by the drivers attached to this usb bus. As a
result I see a uvm panic on suspend in urtwn(4) which is triggered by a
bad rx eof.

OK kettenis@