clockintr: remove clockintr_init(), clockintr_flags
All the state initialization once done in clockintr_init() has been
moved to other parts of the kernel.  It's a dead function.  Remove it.

Likewise, the clockintr_flags variable no longer sports any meaningful
flags.  Remove it.  This frees up the CL_* flag namespace, which might
be useful to the clockintr frontend if we ever need to add behavior
flags to any of those functions.