avoid using MCS9 in iwm(4) Tx rate selection if 40 MHz is disabled
MCS9 is not valid on 20HMz so we must not use it in 20MHz-only mode.
Fixes firmware panic 0x00002078, which indicates the driver sets a rate
with a bad number of antennas, bad number of Tx streams, or bad bandwidth.

Observed by dlg@ with the miktrotik AP for which iwm(4) firmware requires
our 20MHz-only workaround to interoperate in 11ac mode.

ok dlg@