Fix slaac on P2P interfaces
slaacd(8) can work on P2P interfaces, it will just never configure the
destination address. But this works fine on at least pppoe(4) and
tun(4).

To make this less confusing pull ifra_dstaddr into dst6 or gw6
depending on if we are doing autoconf or not.

I accidentally broke this when implementing rule 5.5 of RFC 6724.

reported by & testing naddy
OK bluhm