Fix merge of bounce buffer segments in amd64 bus dma.
If the physical pages are contiguous, _bus_dmamap_load_buffer()
tries to merge the segments.  In case of mbuf chains, it can happen
that the physical bounce buffers are contiguous, but the virtual
addresses of mbuf m_data are not.  Then during transmit _bus_dmamap_sync()
tries to copy segments where it cannot access the virtual source
address which is mapped in a different mbuf.  So if bounce buffers
are used, physical and virtual buffer must be contigous, to merge
a segment.
While there, split check and decrement of variable i in a for loop
to make the code readable.

with and OK hshoexer@