Do not crash in witness code if interrupt fires early.
It can happen that an interrupt handler is called immediately on a
secondary CPU when it is spun up during boot.  At that moment no
process is schduled yet, so curproc is NULL.  To prevent a crash
when p->p_sleeplocks is dereferenced, skip witness check in this
case.

OK visa@